Fix: Sometimes sub-groups don't render in group details mode.
This error would occur in the following situation:

rowval #1: A > B > C > [rows]
rowval #2: A > D > C > [rows]

When rendering the <TR> elements for rowval #2, it would skip "A" (same as previous rowval), then
render "D", then skip "C" (same as previous rowval) - this would create a hierarchy like this:

  A
  +--- B
  |    +--- C
  |
  +--- D

Except that nothing would be put under "D", causing the rows belonging to rowval #2 to become
inaccessible through the details browser.  The fix is to force rendering of all rowval elements once
the first one has been rendered.  So in this case, the fact that we render "D" forces us to render
"C" right after it.