Fix: Sorting by a field that contains numbers after refreshing causes the numbers to be sorted alphabetically.
This was happening because refreshing was not causing the View to
discard its copy of typeInfo.  While View#clearCache() was throwing away
the existing data, it was not throwing away typeInfo.  So later on, when
we do getData() -> getTypeInfo() it was reusing its old typeInfo, which
indicated that the values from the source had already been converted
from strings into numbers.  In this new data, they had not, but since
the internalType was "primitive" the standard <= operator was used for
comparing values in sort(), which works for numbers and strings, and
that's why they were coming out sorted alphabetically.

On not discarding typeInfo, we were doing that up until f21175f1, but
apparently that was removed because it caused an error when loading
prefs with a filter on a column that didn't exist.  Putting this back no
longer causes an error, I'm guessing it was because of the ouroboros of
loading prefs makes the filter UI in the grid control which needs the
typeInfo to know what kind of widget to make, but this happens before
the data has been retrieved, so the typeInfo isn't known.  Well it
doesn't cause an error anymore, but it doesn't look good.