Fix: Non-final rowval elements w/ custom renderers sometimes don't show up.
When in group summary or pivot output, if a rowval element is
represented by a cell with a custom renderer function, and it produces
an Element, that Element would be used on multiple rows, effectively
disappearing from all but the last row.

| Rank    | Last Name |
| ------- | --------- |
|         | Janeway   |
|         | Kirk      |
| Captain | Picard    |

Here, "Captain" is a single Element produced by calling the `render`
function of the cell representing the "Captain" rowval element once.  It
is cached and reused by each row, and since an Element can only appear
on the page once, it basically gets moved down over and over again.

Now the results of the custom render function aren't cached.  In the
example above, the cell representing the "Captain" rowval element has
its render function called three times, instead of just once, producing
three distinct Elements.

| Rank    | Last Name |
| ------- | --------- |
| Captain | Janeway   |
| Captain | Kirk      |
| Captain | Picard    |