Fix: Error when using DataVis that doesn't load immediately.
This was caused by a previous fix that sets the selection immediately
after calling `Grid#render()`.  The problem is that
`Grid#setSelection()` delegates to the grid's renderer, but since
rendering is asynchronous, the renderer isn't guaranteed to be assigned
yet, so the delegation will fail.  The solution is to make the call to
set the selection in the finish callback of the renderer.