Dev: jQuery and jQuery UI are now bundled with DataVis.
jQuery is temporarily bound to `window.jQuery` while jQuery UI is loading, but after that the global
binding is removed.  This means that all JS dependencies are now included in wcdatavis.js without
conflicting with any versions of jQuery that might be loaded afterwards.  However, any versions of
jQuery loaded before... we can't do anything about.